changed ints to doubles

+++ program.c
@@ -3,10 +3,10 @@
 
 int main()
 {
-  int c = 34;
-  int f = c * 9 / 5 + 32;
+  double c = 34;
+  double f = c * 9 / 5 + 32;
 
-  printf("%d C = %d F\n", c, f);
+  printf("%f C = %f F\n", c, f);
 
   return 0;
 }
